First Commit - HTML stuff added
This creates the html folder, which contains a very rough (but workable)
system that will allow you to search for the ticker, scores, and sort by
both of them

I had to use JS ajax for this to function, which means this will only work
in a server setting. I used basic HTML and JS, so it should work right
out the box. A simple HTTP server should do, and point right to the folder
to index.HTML to see it working in action

If a demo is needed, just let me know :)